temp-e1f0b33/requirements.txt
pandas
nltk
scikit-learn
matplotlib
seaborn
Flask
python-dotenv
flask-sqlalchemy
flask-jwt-extended
transformers
tensorflow
tf-keras
hf_xet

flask-cors
requests
torch
python-jose[cryptography] 
werkzeug
python-multipart
fastapi[standard]